Extract ParseImage to internal module
