+++ plugins/inputs/docker/docker_test.go
@@ -8,10 +8,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/influxdata/telegraf/testutil"
-
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/swarm"
+	"github.com/influxdata/telegraf/testutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -812,57 +811,6 @@ func TestContainerName(t *testing.T) {
 					require.Equal(t, tt.expected, metric.Tags["container_name"])
 				}
 			}
-		})
-	}
-}
-
-func TestParseImage(t *testing.T) {
-	tests := []struct {
-		image         string
-		parsedName    string
-		parsedVersion string
-	}{
-		{
-			image:         "postgres",
-			parsedName:    "postgres",
-			parsedVersion: "unknown",
-		},
-		{
-			image:         "postgres:latest",
-			parsedName:    "postgres",
-			parsedVersion: "latest",
-		},
-		{
-			image:         "coreos/etcd",
-			parsedName:    "coreos/etcd",
-			parsedVersion: "unknown",
-		},
-		{
-			image:         "coreos/etcd:latest",
-			parsedName:    "coreos/etcd",
-			parsedVersion: "latest",
-		},
-		{
-			image:         "quay.io/postgres",
-			parsedName:    "quay.io/postgres",
-			parsedVersion: "unknown",
-		},
-		{
-			image:         "quay.io:4443/coreos/etcd",
-			parsedName:    "quay.io:4443/coreos/etcd",
-			parsedVersion: "unknown",
-		},
-		{
-			image:         "quay.io:4443/coreos/etcd:latest",
-			parsedName:    "quay.io:4443/coreos/etcd",
-			parsedVersion: "latest",
-		},
-	}
-	for _, tt := range tests {
-		t.Run("parse name "+tt.image, func(t *testing.T) {
-			imageName, imageVersion := parseImage(tt.image)
-			require.Equal(t, tt.parsedName, imageName)
-			require.Equal(t, tt.parsedVersion, imageVersion)
 		})
 	}
 }
